test(desktop): isolate Node server lifecycle coverage
